src/mjlab/scripts/deploy_template.py
import time
import argparse
import numpy as np
import onnxruntime as ort

# Configure these to match your training config
DECIMATION = 4  # e.g. 200Hz sim / 4 = 50Hz control
DT = 0.02       # 1/50Hz
NUM_ACTIONS = 29
NUM_OBS = 160   # observation size

# Scales from G1 config (verify these!)
ACTION_SCALE = 0.25 
LIN_VEL_SCALE = 2.0
ANG_VEL_SCALE = 0.25
DOF_POS_SCALE = 1.0
DOF_VEL_SCALE = 0.05

class G1RobotInterface:
    def __init__(self):
        # TODO: Initialize Unitree SDK (LowLevel)
        pass

    def get_observation(self):
        """Read sensors from robot."""
        # TODO: Replace with real SDK calls
        
        # Example dummy data
        q = np.zeros(NUM_ACTIONS)      # Joint positions
        dq = np.zeros(NUM_ACTIONS)     # Joint velocities
        quat = np.array([1, 0, 0, 0])  # Base orientation (w, x, y, z)
        omega = np.zeros(3)            # Base angular velocity
        
        return q, dq, quat, omega

    def set_command(self, q_des):
        """Send commands to motors."""
        # TODO: Replace with real SDK calls
        # q_des is target joint positions (PD control usually handled on robot/driver)
        pass

def main(model_path):
    print(f"Loading model from {model_path}...")
    ort_sess = ort.InferenceSession(model_path)

    robot = G1RobotInterface()
    
    # Initialize buffers
    stop_state = False
    default_dof_pos = np.zeros(NUM_ACTIONS) # TODO: Set your default/standing pose

    print("Starting control loop...")
    try:
        while not stop_state:
            start_time = time.perf_counter()

            # 1. Read Sensors
            q, dq, quat, omega = robot.get_observation()

            # 2. Construct Observation Vector
            # NOTE: This must MATCH your training observation order EXACTLY!
            # Standard order: [command, base_lin_vel, base_ang_vel, projected_gravity, joint_pos - default, joint_vel, actions]
            # You might need to estimate base linear velocity if you don't have it (or set to 0)
            
            obs = np.zeros(NUM_OBS, dtype=np.float32)
            # ... fill obs ...

            # 3. Model Inference
            # ONNX runtime expects batch dimension
            inputs = {ort_sess.get_inputs()[0].name: obs[None, :]}
            actions = ort_sess.run(None, inputs)[0][0]

            # 4. Process Actions
            # action = target_pos_delta scaled
            q_target = default_dof_pos + actions * ACTION_SCALE
            
            # 5. Send Command
            robot.set_command(q_target)

            # Frequency Control
            elapsed = time.perf_counter() - start_time
            if elapsed < DT:
                time.sleep(DT - elapsed)

    except KeyboardInterrupt:
        print("Stopping...")

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("model", help="Path to ONNX model file")
    args = parser.parse_args()
    main(args.model)
